+++ src/client/components/pr_scoped_accounts_panel/pr_scoped_accounts_panel.js
@@ -172,7 +172,7 @@ Template.pr_scoped_accounts_panel.helpers({
     parmsTree(){
         return {
             roles: Template.instance().PR.userRoles,
-            pr_edit: false,
+            pr_editable: false,
             pr_div: Template.instance().PR.scoped_div+( Template.instance().PR.editMode.get() ? '-edit' : '-view' ),
             pr_selectable: Boolean( this.pr_selectable !== false ),
             pr_multiple: Boolean( this.pr_selectable === true ),