+++ ChangeLog.md
@@ -10,6 +10,7 @@
     - Fix 'pwix_roles_count_by_roles' publication when an account has redondant assignments
     - Fix the stylesheet to make the tree scrollable
     - Allow multiple account assignements selection
+    - No more allow checboxes changes on view mode
 
 ### 1.6.2